requirements.txt
Flask
flask-cors
gunicorn
numpy
Pillow
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.11.0+cpu